pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools",
]

[project]
name = "metapackages"
description = "Builder script for distribution repository with metapackages"
readme = { file = "README.md", content-type = "text/markdown" }
keywords = [
  "distribution",
  "packages",
]
license = "Apache-2.0"
license-files = [ "LICENSE.txt" ]
maintainers = [
  { name = "Alberto Donato", email = "alberto@albertodonato.net" },
]
authors = [
  { name = "Alberto Donato", email = "alberto@albertodonato.net" },
]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dynamic = [
  "version",
]
dependencies = [
  "click",
  "pyyaml",
]
urls."Issue Tracker" = "https://github.com/albertodonato/metapackages/issues"
urls."Source Code" = "https://github.com/albertodonato/metapackages"
urls.Documentation = "https://github.com/albertodonato/metapackages"
urls.Homepage = "https://github.com/albertodonato/metapackages"
scripts.metapackages = "metapackages:main"

[dependency-groups]
dev = [
  "pre-commit",
  "pyproject-fmt",
  "ruff",
  "ty",
]
types = [
  "types-pyyaml",
]

[tool.setuptools]
dynamic.version = { attr = "metapackages.__version__" }
package-data.metapackages = [ "py.typed" ]
packages.find.include = [ "metapackages*" ]

[tool.ruff]
line-length = 79
lint.select = [ "F", "I", "RUF", "UP" ]
lint.isort.combine-as-imports = true
lint.isort.force-sort-within-sections = true

[tool.pyproject-fmt]
expand_tables = [ "tool.tox" ]

[tool.ty]
rules.unresolved-import = "ignore"

[tool.tox]
env_list = [ "lint", "check" ]

[tool.tox.env.check]
dependency_groups = [ "dev" ]
commands = [ [ "ty", "check", "metapackages", { replace = "posargs", default = [], extend = true } ] ]

[tool.tox.env.lint]
package = "skip"
dependency_groups = [ "dev" ]
commands = [ [ "pre-commit", "run", "--all-files" ] ]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
package = "editable"
